wip #21 refactored Conformer class into analyze.py
